Split MCP dispatcher test fakes into separate namespaces
