+++ tools/CoopMcpServer.Tests/LiveTestCommandDispatcherTests.cs
@@ -51,31 +51,3 @@ namespace CoopMcpServer.Tests
         }
     }
 }
-
-// These fakes exercise the linked dispatcher without loading Bannerlord or native DLLs.
-namespace Common
-{
-    public static class GameThread
-    {
-        public static void Run(Action action, bool blocking) => action();
-    }
-}
-
-namespace TaleWorlds.Library
-{
-    public static class CommandLineFunctionality
-    {
-        private static readonly Dictionary<string, object> AllFunctions = new()
-        {
-            ["coop.debug.legacy"] = new(), ["campaign.do_something"] = new(), ["coop.unregistered"] = new(),
-        };
-        public static void CollectCommandLineFunctions() { }
-        public static string CallFunction(string command, List<string> arguments, out bool found)
-        {
-            if (!command.StartsWith("coop.debug.", StringComparison.Ordinal))
-                throw new InvalidOperationException("Arbitrary vanilla dispatch is forbidden in this test.");
-            found = AllFunctions.ContainsKey(command);
-            return "legacy result";
-        }
-    }
-}
